# server/docs/swagger.yaml
definitions:
  request.Register:
    properties:
      authorityId:
        example: int 角色id
        type: string
      authorityIds:
        example: '[]uint 角色id'
        type: string
      email:
        example: 电子邮箱
        type: string
      enable:
        example: int 是否启用
        type: string
      password:
        example: 密码
        type: string
      phone:
        example: 电话号码
        type: string
      username:
        example: 用户名
        type: string
    required:
    - password
    - username
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  response.SysUserResponse:
    properties:
      user:
        $ref: '#/definitions/system.SysUser'
    type: object
  system.Meta:
    properties:
      activeName:
        type: string
      closeTab:
        type: boolean
      icon:
        type: string
      keepAlive:
        type: boolean
      title:
        type: string
    type: object
  system.SysAuthority:
    properties:
      authorityId:
        description: 角色ID
        type: integer
      authorityName:
        description: 角色名
        type: string
      children:
        description: 子角色会在业务中动态读取，不写入表中
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      createdAt:
        description: 创建时间
        type: string
      dataAuthorityId:
        description: 角色之间多对多关系，gorm 自动创建 sys_data_authority_id 映射表
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      defaultRouter:
        description: 默认菜单(默认dashboard)
        type: string
      deletedAt:
        type: string
      menus:
        items:
          $ref: '#/definitions/system.SysBaseMenu'
        type: array
      parentId:
        description: 父角色ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenu:
    properties:
      ID:
        type: integer
      authorities:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      children:
        items:
          $ref: '#/definitions/system.SysBaseMenu'
        type: array
      component:
        type: string
      createdAt:
        type: string
      hidden:
        type: boolean
      menuBtn:
        items:
          $ref: '#/definitions/system.SysBaseMenuBtn'
        type: array
      meta:
        allOf:
        - $ref: '#/definitions/system.Meta'
        description: 把 Meta 中的字段嵌入到该表中，不创建新的关联表
      name:
        type: string
      parameters:
        items:
          $ref: '#/definitions/system.SysBaseMenuParameter'
        type: array
      parentId:
        type: integer
      path:
        type: string
      sort:
        type: integer
      updatedAt:
        type: string
    type: object
  system.SysBaseMenuBtn:
    properties:
      ID:
        type: integer
      createdAt:
        type: string
      desc:
        type: string
      name:
        type: string
      sysBaseMenuID:
        type: integer
      updatedAt:
        type: string
    type: object
  system.SysBaseMenuParameter:
    properties:
      ID:
        type: integer
      createdAt:
        type: string
      key:
        type: string
      sysBaseMenuID:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
      value:
        type: string
    type: object
  system.SysUser:
    properties:
      ID:
        type: integer
      authorities:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      authority:
        $ref: '#/definitions/system.SysAuthority'
      authorityId:
        type: integer
      createdAt:
        type: string
      email:
        type: string
      enable:
        type: integer
      phone:
        type: string
      updatedAt:
        type: string
      username:
        type: string
      uuid:
        type: string
    type: object
info:
  contact: {}
  description: gin+vue的BI系统
  title: BIGO Swagger API接口文档
  version: v1.0.0
paths:
  /user/admin_register:
    post:
      parameters:
      - description: 用户名, 密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Register'
      produces:
      - application/json
      responses:
        "200":
          description: 注册用户返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.SysUserResponse'
                msg:
                  type: string
              type: object
      summary: 用户注册
      tags:
      - sysUser
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"