# server/model/system/request/sysUser.go
package request

type Register struct {
	Username     string `json:"username" binding:"required" example:"用户名"`
	Password     string `json:"password" binding:"required" example:"密码"`
	AuthorityId  uint   `json:"authorityId" swaggertype:"string" example:"int 角色id"`
	Enable       int    `json:"enable" swaggertype:"string" example:"int 是否启用"`
	AuthorityIds []uint `json:"authorityIds" swaggertype:"string" example:"[]uint 角色id"`
	Phone        string `json:"phone" example:"电话号码"`
	Email        string `json:"email" example:"电子邮箱"`
}